Allow build of modules for not-running kernel versions
The current Makefile always sets KVER and some other variables to the
version of the currently running kernel and does not allow overriding
this value. So installing a newer kernel and building and installing
the modules for the new version will not work before activation of the
new kernel. With this patch, the following can be done:

export KVER=<newly-installed-kernel-version>
echo "Kernel version to build for: $KVER"
export MOD_DIR=/lib/modules/${KVER}/extra/
KVER=${KVER} MOD_DIR=${MOD_DIR} HDAPS=1 make

Signed-off-by: Joerg Mayer <jmayer@loplof.de>